unsigned int

+++ Cplus/ContainerWithMostWater.cpp
@@ -1,5 +1,5 @@
-#include <vector>
 #include <algorithm>
+#include <vector>
 using namespace std;
 
 class Solution
@@ -8,7 +8,7 @@ public:
 	int maxArea(vector<int> &height)
 	{
 		int res = 0;
-		for (int i = 0, j = height.size() - 1; i < j;)
+		for (int i = 0, j = (int)height.size() - 1; i < j;)
 		{
 			res = max(res, (j - i) * min(height[i], height[j]));
 			height[i] < height[j] ? ++i : --j;
